fix: timestamp Integer -> double 로 타입 변경

+++ src/main/java/com/swmStrong/demo/domain/leaderboard/service/LeaderboardService.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface LeaderboardService {
-    void increaseScore(String categoryId, String userId, double duration, Integer timestamp);
+    void increaseScore(String categoryId, String userId, double duration, double timestamp);
     List<LeaderboardResponseDto> getLeaderboardPage(String category, int page, int size, LocalDate date, PeriodType periodType);
     LeaderboardResponseDto getUserScoreInfo(String category, String userId, LocalDate date, PeriodType periodType);
     List<LeaderboardResponseDto> getAllLeaderboard(String category, LocalDate date);
