+++ src/main/java/com/swmStrong/demo/domain/leaderboard/service/LeaderboardServiceImpl.java
@@ -37,8 +37,8 @@ public class LeaderboardServiceImpl implements LeaderboardService {
     }
 
     @Override
-    public void increaseScore(String categoryId, String userId, double duration, Integer timestamp) {
-        LocalDate day = LocalDateTime.ofInstant(java.time.Instant.ofEpochSecond(timestamp),
+    public void increaseScore(String categoryId, String userId, double duration, double timestamp) {
+        LocalDate day = LocalDateTime.ofInstant(java.time.Instant.ofEpochSecond((long) timestamp),
                 java.time.ZoneId.systemDefault()).toLocalDate();
         ObjectId categoryObjectId = new ObjectId(categoryId);
         String category = categoryProvider.getCategoryById(categoryObjectId);
